Adding passing Alarm tests, adding failing service tests
